+++ CHANGELOG.md
@@ -6,6 +6,10 @@ in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Change glyph into an Option, so that we can handle cards without a unicode character to match.
+
 ## [0.8.5] - 2019-01-13
 ### Added
 - Adding support for drawing a card from a deck of cards.